requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# txt/conti_wc.py: 6
Pillow == 8.4.0

# txt/conti_wc.py: 10
matplotlib == 3.5.1

# txt/conti_wc.py: 8
nltk == 3.6.3

# txt/conti_wc.py: 9
numpy == 1.20.3

# txt/app.py: 6
# txt/graph.py: 4
pandas == 1.3.5

# txt/graph.py: 5
plotly == 5.5.0

# txt/conti_wc.py: 12
scipy == 1.7.3

# txt/conti_wc.py: 13
wordcloud == 1.8.1
